requirements.txt
langchain_community
bs4
langchain-ollama
urllib3
pypdf
chromadb
python-dotenv